requirements.txt
flask
langchain
langchain-community
langchain-core
langchain-text-splitters
chromadb
pandas
python-dotenv
sentence-transformers
openai
tiktoken
gunicorn